[#757] Split lexical-editor.tsx into smaller modules
Refactored the 1646-line lexical-editor.tsx into a modular structure
for improved maintainability. No functional changes.

New structure:
- types.ts - TypeScript interfaces and types
- config/theme.ts - Editor theme configuration
- utils/url-validation.ts - URL validation for links
- utils/sanitize.ts - DOMPurify HTML sanitization
- utils/markdown-to-html.ts - Markdown to HTML conversion
- utils/highlight.ts - Code syntax highlighting with lazy loading
- components/toolbar-button.tsx - Toolbar button component
- components/toolbar-separator.tsx - Toolbar separator component
- components/mermaid-renderer.tsx - Mermaid diagram rendering
- dialogs/link-dialog.tsx - Link insertion dialog
- dialogs/table-dialog.tsx - Table insertion dialog
- plugins/toolbar-plugin.tsx - Formatting toolbar plugin
- plugins/initial-content-plugin.tsx - Initial content loader
- plugins/content-sync-plugin.tsx - Content sync and save shortcut
- plugins/auto-focus-plugin.tsx - Auto-focus plugin
- plugins/code-highlight-plugin.tsx - Code highlighting plugin
- lexical-editor.tsx - Main editor component (now ~300 lines)
- index.ts - Updated exports for all modules
