+++ src/ui/components/notes/editor/components/toolbar-separator.tsx
@@ -1,0 +1,10 @@
+/**
+ * Toolbar separator component for the editor.
+ * Part of Epic #338, Issue #757
+ */
+
+import React from 'react';
+
+export function ToolbarSeparator(): React.JSX.Element {
+  return <div className="w-px h-6 bg-border mx-1" />;
+}
