+++ src/ui/components/notes/editor/dialogs/link-dialog.tsx
@@ -1,0 +1,107 @@
+/**
+ * Link insertion dialog component.
+ * Part of Epic #338, Issue #757
+ *
+ * Addresses issues #675 (replace prompt()) and #678 (URL validation).
+ */
+
+import React, { useState, useEffect, useRef } from 'react';
+import {
+  Dialog,
+  DialogContent,
+  DialogDescription,
+  DialogFooter,
+  DialogHeader,
+  DialogTitle,
+} from '@/ui/components/ui/dialog';
+import { Input } from '@/ui/components/ui/input';
+import { Label } from '@/ui/components/ui/label';
+import { Button } from '@/ui/components/ui/button';
+import { validateUrl, normalizeUrl } from '../utils/url-validation';
+import type { LinkDialogProps } from '../types';
+
+export function LinkDialog({
+  open,
+  onOpenChange,
+  onSubmit,
+}: LinkDialogProps): React.JSX.Element {
+  const [url, setUrl] = useState('');
+  const [error, setError] = useState<string | null>(null);
+  const inputRef = useRef<HTMLInputElement>(null);
+
+  // Focus input when dialog opens
+  useEffect(() => {
+    if (open) {
+      setUrl('');
+      setError(null);
+      // Small delay to ensure dialog is mounted
+      setTimeout(() => inputRef.current?.focus(), 50);
+    }
+  }, [open]);
+
+  const handleSubmit = (e: React.FormEvent) => {
+    e.preventDefault();
+
+    const validationError = validateUrl(url);
+    if (validationError) {
+      setError(validationError);
+      return;
+    }
+
+    const normalizedUrl = normalizeUrl(url);
+    onSubmit(normalizedUrl);
+    onOpenChange(false);
+  };
+
+  const handleKeyDown = (e: React.KeyboardEvent) => {
+    if (e.key === 'Enter') {
+      e.preventDefault();
+      handleSubmit(e);
+    }
+  };
+
+  return (
+    <Dialog open={open} onOpenChange={onOpenChange}>
+      <DialogContent className="sm:max-w-md">
+        <DialogHeader>
+          <DialogTitle>Insert Link</DialogTitle>
+          <DialogDescription>
+            Enter a URL to create a link. Supported protocols: http, https, mailto, tel.
+          </DialogDescription>
+        </DialogHeader>
+        <form onSubmit={handleSubmit}>
+          <div className="grid gap-4 py-4">
+            <div className="grid gap-2">
+              <Label htmlFor="link-url">URL</Label>
+              <Input
+                id="link-url"
+                ref={inputRef}
+                type="text"
+                value={url}
+                onChange={(e) => {
+                  setUrl(e.target.value);
+                  setError(null);
+                }}
+                onKeyDown={handleKeyDown}
+                placeholder="https://example.com"
+                aria-invalid={error ? 'true' : 'false'}
+                aria-describedby={error ? 'link-url-error' : undefined}
+              />
+              {error && (
+                <p id="link-url-error" className="text-sm text-destructive">
+                  {error}
+                </p>
+              )}
+            </div>
+          </div>
+          <DialogFooter>
+            <Button type="button" variant="outline" onClick={() => onOpenChange(false)}>
+              Cancel
+            </Button>
+            <Button type="submit">Insert Link</Button>
+          </DialogFooter>
+        </form>
+      </DialogContent>
+    </Dialog>
+  );
+}
