+++ src/ui/components/notes/editor/plugins/auto-focus-plugin.tsx
@@ -1,0 +1,17 @@
+/**
+ * Plugin to auto-focus the editor.
+ * Part of Epic #338, Issue #757
+ */
+
+import { useEffect } from 'react';
+import { useLexicalComposerContext } from '@lexical/react/LexicalComposerContext';
+
+export function AutoFocusPlugin(): null {
+  const [editor] = useLexicalComposerContext();
+
+  useEffect(() => {
+    editor.focus();
+  }, [editor]);
+
+  return null;
+}
