+++ src/ui/components/notes/editor/plugins/code-highlight-plugin.tsx
@@ -1,0 +1,18 @@
+/**
+ * Plugin to enable code highlighting in code blocks.
+ * Part of Epic #338, Issue #757
+ */
+
+import { useEffect } from 'react';
+import { useLexicalComposerContext } from '@lexical/react/LexicalComposerContext';
+import { registerCodeHighlighting } from '@lexical/code';
+
+export function CodeHighlightPlugin(): null {
+  const [editor] = useLexicalComposerContext();
+
+  useEffect(() => {
+    return registerCodeHighlighting(editor);
+  }, [editor]);
+
+  return null;
+}
